+++ Foundation/Auth/System/Flow/OAuth/IntrospectToken/IntrospectTokenData.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Avax\Auth\System\Flow\OAuth\IntrospectToken;
+
+use SensitiveParameter;
+
+final readonly class IntrospectTokenData
+{
+    public function __construct(
+        public string $clientId,
+        #[SensitiveParameter] public string $token,
+        #[SensitiveParameter] public string|null $clientSecret = null
+    ) {}
+}