+++ Foundation/Auth/System/Flow/OAuth/RevokeToken/RevokeTokenData.php
@@ -1,0 +1,19 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Avax\Auth\System\Flow\OAuth\RevokeToken;
+
+use SensitiveParameter;
+
+final readonly class RevokeTokenData
+{
+    public function __construct(
+        public string $clientId,
+        #[SensitiveParameter] public string $token,
+        #[SensitiveParameter] public string|null $clientSecret = null,
+        public string|null $tokenTypeHint = null,
+        public string|null $ipAddress = null,
+        public string|null $userAgent = null
+    ) {}
+}